feat(api): update user management endpoints and improve error handling; add FoodieTraveler user routes
