+++ docker-compose.yml
@@ -11,7 +11,7 @@ services:
     ports:
       - "5432:5432"
     volumes:
-      - neon_data:/var/lib/postgresql/data
+      - neon_data1:/var/lib/postgresql/data
   neon-db2:
     image: postgres:15
     container_name: neon-postgres-2
@@ -23,7 +23,8 @@ services:
     ports:
       - "5433:5432"
     volumes:
-      - neon_data:/var/lib/postgresql/data
+      - neon_data2:/var/lib/postgresql/data
 
 volumes:
-  neon_data:
+  neon_data1:
+  neon_data2:
