+++ src/app/api/FoodieTraveler/users/[id]/route.ts
@@ -1,0 +1,66 @@
+import { NextRequest, NextResponse } from "next/server";
+import { PrismaClient } from "@/../prisma/generated/postgresql";
+
+const prisma = new PrismaClient();
+// ✅ Get a User by ID (GET)
+export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
+  try {
+    const { id } = params;
+
+    const user = await prisma.userFoodie.findUnique({
+      where: { id },
+      select: { id: true, email: true, name: true, createdAt: true },
+    });
+
+    if (!user) {
+      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    }
+
+    return NextResponse.json(user, { status: 200 });
+
+  } catch (error) {
+    console.error("Error fetching user:", error);
+    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
+  }
+}
+
+// ✅ Update User Details (PUT)
+export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
+  try {
+    const { id } = params;
+    const { name } = await req.json();
+
+    if (!name) {
+      return NextResponse.json({ error: "Name is required" }, { status: 400 });
+    }
+
+    const updatedUser = await prisma.userFoodie.update({
+      where: { id },
+      data: { name },
+      select: { id: true, email: true, name: true },
+    });
+
+    return NextResponse.json(updatedUser, { status: 200 });
+
+  } catch (error) {
+    console.error("Error updating user:", error);
+    return NextResponse.json({ error: "Failed to update user" }, { status: 500 });
+  }
+}
+
+// ✅ Ban (Delete) a User (DELETE)
+export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
+  try {
+    const { id } = params;
+
+    await prisma.userFoodie.delete({
+      where: { id },
+    });
+
+    return NextResponse.json({ message: "User banned successfully" }, { status: 200 });
+
+  } catch (error) {
+    console.error("Error banning user:", error);
+    return NextResponse.json({ error: "Failed to ban user" }, { status: 500 });
+  }
+}
